Trying to fix remaining leaks and some issues writing to file

+++ indexer/src/index.h
@@ -4,11 +4,10 @@
 #include "../../util/hashtable.h"
 #include "../../util/common.h"
 #include "../../util/list.h"
-#include <pthread.h>
 
 /* Wordnode Hashtable wrappers */
-int updateIndex(char* word, intptr_t docID, HashTable* index);
-int initWNode(char* word, intptr_t docID, WordNode* wNode);
+int updateIndex(char* word, int docID, HashTable* index);
+int initWNode(char* word, int docID, WordNode* wNode);
 void dNode_free(element_t elem);
 int dNode_cmp(element_t av, element_t bv);
 void IndexLoadWords(HashTable* Index, char** str);
